fix: feedProfileViewResponseSchema를 서버 참조 가능한 파일로 분리
